+++ Assets/Hashimoto/Prefab/ブラックサンタ（仮）.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: c12b0fd07e4e5c64f8e5365c114b6776
-folderAsset: yes
-DefaultImporter:
+guid: 8ddaeac8f360a4941a3111f12ec55576
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
