+++ Assets/Hashimoto/Scripts/Enemy_Capsule.cs
@@ -8,20 +8,26 @@ public class Enemy_Capsule : MonoBehaviour
     [Header("ブラックサンタの体力")]
     [SerializeField] float _hp = 10;
 
-    public float HP
+    public float EnemyHP
     {
         get
         {
             return _hp;
         }
     }
-    
-    
+
+    private void Update()
+    {
+        print(_hp);
+    }
+
+
     private void OnTriggerEnter2D(Collider2D collision)
     {
         SnowBall YukiObj = collision.GetComponent<SnowBall>();
         if (YukiObj && YukiObj.Type == TargetType.Enemy)
         {
+            print("a");
             _hp -= YukiObj.Damage;
         }
     }
